requirements.txt
# python<=3.9
black==21.11b1
click==8.0.3
GitPython==3.1.24
pybars4==0.9.13
pygments>=2.10
